+++ src/ipc/utils/app_upgrade_utils.ts
@@ -38,29 +38,6 @@ export function isComponentTaggerUpgradeNeeded(appPath: string): boolean {
   }
 }
 
-export function isCapacitorUpgradeNeeded(appPath: string): boolean {
-  // Check if it's a Vite app first
-  if (!isViteApp(appPath)) {
-    return false;
-  }
-
-  // Check if Capacitor is already installed
-  const capacitorConfigJs = path.join(appPath, "capacitor.config.js");
-  const capacitorConfigTs = path.join(appPath, "capacitor.config.ts");
-  const capacitorConfigJson = path.join(appPath, "capacitor.config.json");
-
-  // If any Capacitor config exists, the upgrade is not needed
-  if (
-    fs.existsSync(capacitorConfigJs) ||
-    fs.existsSync(capacitorConfigTs) ||
-    fs.existsSync(capacitorConfigJson)
-  ) {
-    return false;
-  }
-
-  return true;
-}
-
 export async function applyComponentTagger(appPath: string) {
   const viteConfigPathJs = path.join(appPath, "vite.config.js");
   const viteConfigPathTs = path.join(appPath, "vite.config.ts");
@@ -112,7 +89,7 @@ export async function applyComponentTagger(appPath: string) {
     }
   } else {
     throw new Error(
-      "Could not find `plugins: [` in vite.config.ts. Manual installation required.",
+      `Could not find 'plugins: [' in ${path.basename(viteConfigPath)}. Manual installation required.`,
     );
   }
 
@@ -162,76 +139,6 @@ export async function applyComponentTagger(appPath: string) {
   } catch (err) {
     logger.warn(
       `Failed to commit changes. This may happen if the project is not in a git repository, or if there are no changes to commit.`,
-      err,
-    );
-  }
-}
-
-export async function applyCapacitor({
-  appName,
-  appPath,
-}: {
-  appName: string;
-  appPath: string;
-}) {
-  const env = getPackageManagerCommandEnv() as Record<string, string>;
-
-  // Install Capacitor dependencies
-  await simpleSpawn({
-    command:
-      "pnpm add --ignore-workspace-root-check @capacitor/core@7.4.4 @capacitor/cli@7.4.4 @capacitor/ios@7.4.4 @capacitor/android@7.4.4 || npm install @capacitor/core@7.4.4 @capacitor/cli@7.4.4 @capacitor/ios@7.4.4 @capacitor/android@7.4.4 --legacy-peer-deps",
-    cwd: appPath,
-    env,
-    successMessage: "Capacitor dependencies installed successfully",
-    errorPrefix: "Failed to install Capacitor dependencies",
-  });
-
-  // Initialize Capacitor
-  await simpleSpawn({
-    command: `npx cap init "${appName}" "com.example.${appName.toLowerCase().replace(/[^a-z0-9]/g, "")}" --web-dir=dist`,
-    cwd: appPath,
-    env,
-    successMessage: "Capacitor initialized successfully",
-    errorPrefix: "Failed to initialize Capacitor",
-  });
-
-  // Intentionally omit PNPM_INSTALL_POLICY_ARGS because:
-  // 1. confirmModulesPurge will almost never be needed for capacitor (i.e. user would need to switch from npm to pnpm and not triggered a rebuild).
-  // 2. strictBuildDeps should be kept true (default value) in case capacitor has native deps.
-  await simpleSpawn({
-    command:
-      "pnpm install --prod=false || npm install --include=dev --legacy-peer-deps",
-    cwd: appPath,
-    env,
-    successMessage: "Development dependencies installed successfully",
-    errorPrefix: "Failed to install development dependencies",
-  });
-
-  // Add iOS and Android platforms
-  await simpleSpawn({
-    command: "npx cap add ios && npx cap add android",
-    cwd: appPath,
-    env,
-    successMessage: "iOS and Android platforms added successfully",
-    errorPrefix: "Failed to add iOS and Android platforms",
-  });
-
-  // Commit changes
-  try {
-    logger.info("Staging and committing Capacitor changes");
-    await gitAddAll({ path: appPath });
-    await gitCommit({
-      path: appPath,
-      message: "[dyad] add Capacitor for mobile app support",
-    });
-    logger.info("Successfully committed Capacitor changes");
-  } catch (err) {
-    logger.warn(
-      `Failed to commit changes. This may happen if the project is not in a git repository, or if there are no changes to commit.`,
-      err,
-    );
-    throw new Error(
-      "Failed to commit Capacitor changes. Please commit them manually. Error: " +
       err,
     );
   }
